Update GitHub Actions workflow and enhance MyEventLane module dependencies and services. Upgraded download-artifact action to v6, added dependency on myeventlane_core in myeventlane_help_assistant, and introduced event_state_resolver service for improved event handling in EventSuggestionEngine.
